Showing the nodes that are behind the user; and temporary code for Ground Plane on each Slice level
Showing the nodes that are behind the user - for VR experience user
needs to have the ability to look behind him and see the nodes which
previously were not drawn. Previously such nodes were not computed for
better performance.
Also added temporary code for Ground Plane on each Slice level -
commented for now because of unknown future usage.