Image Precision corrected in LoadTexture()
Image Precision corrected in LoadTexture(). It is preferred to actually
give a real image format, one with a specific internal precision. If the
OpenGL implementation does not support the particular format and
precision you choose, the driver will internally convert it into
something it does support.
OpenGL versions 3.x and above have a set of required image formats that
all conforming implementations must implement.
Although GL will accept GL_RGB​, it is up to the driver to decide an
appropriate precision. It is recommended to be specific and write
GL_RGB8​ (GL_RGBA8 in most GPUs):
https://www.opengl.org/wiki/Common_Mistakes